+++ src/main/java/com/minhcraft/beyondbetatweaks/register/BeyondBetaSounds.java
@@ -10,6 +10,7 @@ public class BeyondBetaSounds {
     public static SoundEvent GIANT_DEATH = register("entity.giant.death");
     public static SoundEvent GIANT_HURT = register("entity.giant.hurt");
     public static SoundEvent GIANT_AMBIENT = register("entity.giant.ambient");
+    public static SoundEvent CHICKEN_SHED_FEATHERS = register("entity.chicken.shed_feathers");
 
     public static void init() {
     }